Optimal Subsampling Methods for Generalized Linear Models

Draw subsample ...

ssp.glm(
  formula,
  data,
  subset = NULL,
  n.plt,
  n.ssp,
  family = "binomial",
  criterion = "optL",
  sampling.method = "poisson",
  likelihood = "weighted",
  control = list(...),
  contrasts = NULL,
  ...
)

Arguments

formula

An object of class "formula" which describes the model to be fitted.

data

A data frame containing the variables in the model.

subset

An optional vector specifying a subset of observations to be used.

n.plt

The pilot subsample size (the first-step subsample size). These samples will be used to estimate the pilot estimator as well as to estimate the optimal subsampling probability.

n.ssp

The expectation optimal subsample size (the second-step subsample size). For sampling.method = 'withReplacement', n.ssp is exactly the subsample size. For sampling.method = 'poisson', n.ssp is the expectation of subsample size.

family

family can be a character string naming a family function, a family function or the result of a call to a family function. Currently 'binomial', 'poisson' and 'Gamma' are implemented.

criterion

The criterion of optimal subsampling probabilities. Choices include optA, optL(default), LCC and uniform.

sampling.method

The sampling method for drawing the optimal subsample. Choices include withReplacement and poisson(default).

likelihood

The type of the maximum likelihood function used to calculate the optimal subsampling estimator. Choices include weighted and logOddsCorrection.

control

A list of parameters for controlling the sampling process. Default is list(alpha=0, b=2).

contrasts

An optional list. It specifies how categorical variables are represented in the design matrix. For example, contrasts = list(v1 = 'contr.treatment', v2 = 'contr.sum').

...

A list of parameters which will be passed to svyglm().

Value

ssp.glm returns an object of class "ssp.glm" containing the following components (some are optional):

model.call

model call

beta.plt

pilot estimator

beta.ssp

optimal subsample estimator.

coefficients

weighted combination of beta.plt and beta.ssp.

cov.ssp

covariance matrix of beta.ssp

cov

covariance matrix of beta.cmb

index.plt

index of pilot subsample in the full sample

index.ssp

index of optimal subsample in the full sample

N

number of observations in the full sample

subsample.size.expect

expected subsample size

terms

model terms

Details

As suggested by survey::svyglm(), for binomial and poisson families use family=quasibinomial() and family=quasipoisson() to avoid a warning "In eval(family$initialize) : non-integer #successes in a binomial glm!". The warning is due to the non-integer survey weights. The ‘quasi’ versions of the family objects give the same point estimates and do not give the warning. Subsampling methods only use point estimates from svyglm() for further computation so that would not bring problems. For Gamma family, it will only return the estimation of coefficients, not dispersion parameter.

likelihood = logOddsCorrection is implemented only for logistic regression (family = binomial or quasibonomial).

In control, alpha is the mixture proportions of optimal subsampling probability and uniform sampling probability. b is the parameter controls the upper threshold for optimal subsampling probability.

References

Wang, H. (2019). More efficient estimation for logistic regression with optimal subsamples. Journal of machine learning research, 20(132), 1-59. https://www.jmlr.org/papers/v20/18-596.html

Ai, M., Yu, J., Zhang, H., & Wang, H. (2021). Optimal subsampling algorithms for big data regressions. Statistica Sinica, 31(2), 749-772. doi:10.5705/ss.202018.0439

Wang, H., & Kim, J. K. (2022). Maximum sampled conditional likelihood for informative subsampling. Journal of machine learning research, 23(332), 1-50. https://www.jmlr.org/papers/v23/21-0506.html

Examples

# logistic regression
set.seed(2)
N <- 1e4
beta0 <- rep(-0.5, 7)
d <- length(beta0) - 1
X <- matrix(0, N, d)
generate_rexp <- function(x) x <- rexp(N, rate = 2)
X <- apply(X, 2, generate_rexp)
Y <- rbinom(N, 1, 1 - 1 / (1 + exp(beta0[1] + X %*% beta0[-1])))
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ .
n.plt <- 500
n.ssp <- 1000
subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'quasibinomial',
criterion = "optL",
sampling.method = 'poisson',
likelihood = "logOddsCorrection")
summary(subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "quasibinomial", criterion = "optL", sampling.method = "poisson", 
#>     likelihood = "logOddsCorrection")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size  1000
#> 3   Actual Subsample Size   999
#> 4   Unique Subsample Size   999
#> 5  Expected Subample Rate   10%
#> 6    Actual Subample Rate 9.99%
#> 7    Unique Subample Rate 9.99%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept  -0.2514     0.1558 -1.6130   0.1067
#> V2         -0.7128     0.1246 -5.7186  <0.0001
#> V3         -0.3954     0.1176 -3.3619   0.0008
#> V4         -0.6214     0.1237 -5.0233  <0.0001
#> V5         -0.7741     0.1298 -5.9633  <0.0001
#> V6         -0.5324     0.1222 -4.3553  <0.0001
#> V7         -0.7193     0.1194 -6.0249  <0.0001
subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'binomial', 
criterion = "optL",
sampling.method = 'withReplacement', 
likelihood = "weighted")
#> Warning: non-integer #successes in a binomial glm!
#> Warning: non-integer #successes in a binomial glm!
summary(subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "binomial", criterion = "optL", sampling.method = "withReplacement", 
#>     likelihood = "weighted")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size  1000
#> 3   Actual Subsample Size  1000
#> 4   Unique Subsample Size   907
#> 5  Expected Subample Rate   10%
#> 6    Actual Subample Rate   10%
#> 7    Unique Subample Rate 9.07%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept  -0.5442     0.1601 -3.3987   0.0007
#> V2         -0.4767     0.1302 -3.6605   0.0003
#> V3         -0.4002     0.1270 -3.1511   0.0016
#> V4         -0.4816     0.1354 -3.5581   0.0004
#> V5         -0.5575     0.1241 -4.4913  <0.0001
#> V6         -0.2991     0.1298 -2.3042   0.0212
#> V7         -0.3519     0.1328 -2.6490   0.0081
Uni.subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'binomial', 
criterion = 'uniform')
summary(Uni.subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "binomial", criterion = "uniform")
#> 
#> Subsample Size:
#>                                 
#> 1       Total Sample Size  10000
#> 2 Expected Subsample Size   1500
#> 3   Actual Subsample Size   1501
#> 4   Unique Subsample Size   1501
#> 5  Expected Subample Rate    15%
#> 6    Actual Subample Rate 15.01%
#> 7    Unique Subample Rate 15.01%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept  -0.6398     0.2091 -3.0602   0.0022
#> V2         -0.3599     0.1764 -2.0407   0.0413
#> V3         -0.4884     0.1784 -2.7381   0.0062
#> V4         -0.4454     0.1698 -2.6225   0.0087
#> V5         -0.6423     0.1906 -3.3696   0.0008
#> V6         -0.3582     0.1843 -1.9438   0.0519
#> V7         -0.5134     0.1575 -3.2594   0.0011
####################
# poisson regression
set.seed(1)
N <-  1e4
beta0 <- rep(0.5, 7)
d <- length(beta0) - 1
X <- matrix(runif(N * d), N, d)
epsilon <- runif(N)
lambda <- exp(beta0[1] + X %*% beta0[-1])
Y <- rpois(N, lambda)
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ .
n.plt <- 200
n.ssp <- 600
subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'poisson',
criterion = "optL", 
sampling.method = 'poisson',
likelihood = "weighted")
summary(subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "poisson", criterion = "optL", sampling.method = "poisson", 
#>     likelihood = "weighted")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size   600
#> 3   Actual Subsample Size   681
#> 4   Unique Subsample Size   681
#> 5  Expected Subample Rate    6%
#> 6    Actual Subample Rate 6.81%
#> 7    Unique Subample Rate 6.81%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept   0.5595     0.0570  9.8147  <0.0001
#> V2          0.4908     0.0357 13.7524  <0.0001
#> V3          0.5131     0.0371 13.8469  <0.0001
#> V4          0.4262     0.0352 12.1229  <0.0001
#> V5          0.4990     0.0398 12.5267  <0.0001
#> V6          0.4952     0.0383 12.9155  <0.0001
#> V7          0.4948     0.0378 13.0972  <0.0001
subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'poisson', 
criterion = "optL", 
sampling.method = 'withReplacement',
likelihood = "weighted")
summary(subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "poisson", criterion = "optL", sampling.method = "withReplacement", 
#>     likelihood = "weighted")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size   600
#> 3   Actual Subsample Size   600
#> 4   Unique Subsample Size   568
#> 5  Expected Subample Rate    6%
#> 6    Actual Subample Rate    6%
#> 7    Unique Subample Rate 5.68%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept   0.5543     0.0590  9.3997  <0.0001
#> V2          0.4505     0.0402 11.2074  <0.0001
#> V3          0.4643     0.0423 10.9781  <0.0001
#> V4          0.5392     0.0420 12.8471  <0.0001
#> V5          0.4497     0.0425 10.5933  <0.0001
#> V6          0.4529     0.0415 10.9180  <0.0001
#> V7          0.5495     0.0395 13.9048  <0.0001
Uni.subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'poisson', 
criterion = 'uniform')
summary(Uni.subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "poisson", criterion = "uniform")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size   800
#> 3   Actual Subsample Size   820
#> 4   Unique Subsample Size   820
#> 5  Expected Subample Rate    8%
#> 6    Actual Subample Rate  8.2%
#> 7    Unique Subample Rate  8.2%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept   0.4226     0.0591  7.1479  <0.0001
#> V2          0.5789     0.0431 13.4381  <0.0001
#> V3          0.4852     0.0425 11.4087  <0.0001
#> V4          0.5401     0.0427 12.6439  <0.0001
#> V5          0.4419     0.0432 10.2351  <0.0001
#> V6          0.5156     0.0436 11.8327  <0.0001
#> V7          0.5311     0.0418 12.6998  <0.0001
##################
# gamma regression
set.seed(1)
N <- 1e4
p <- 3
beta0 <- rep(0.5, p + 1)
d <- length(beta0) - 1
shape <- 2
X <- matrix(runif(N * d), N, d)
link_function <- function(X, beta0) 1 / (beta0[1] + X %*% beta0[-1])
scale <- link_function(X, beta0) / shape
Y <- rgamma(N, shape = shape, scale = scale)
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ .
n.plt <- 200
n.ssp <- 1000
subsampling.results <- ssp.glm(formula = formula, 
data = data, 
n.plt = n.plt,
n.ssp = n.ssp,
family = 'Gamma',
criterion = "optL", 
sampling.method = 'poisson',
likelihood = "weighted")
summary(subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.glm(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     family = "Gamma", criterion = "optL", sampling.method = "poisson", 
#>     likelihood = "weighted")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size  1000
#> 3   Actual Subsample Size   998
#> 4   Unique Subsample Size   998
#> 5  Expected Subample Rate   10%
#> 6    Actual Subample Rate 9.98%
#> 7    Unique Subample Rate 9.98%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept   0.5077     0.0507 10.0224  <0.0001
#> V2          0.4790     0.0605  7.9234  <0.0001
#> V3          0.3964     0.0621  6.3839  <0.0001
#> V4          0.4981     0.0678  7.3515  <0.0001