Optimal Subsampling Methods for Quantile Regression Model

This function fits generalized linear models using ....

ssp.quantreg(
  formula,
  data,
  subset = NULL,
  tau = 0.5,
  n.plt,
  n.ssp,
  B = 5,
  boot = TRUE,
  criterion = "optL",
  sampling.method = "withReplacement",
  likelihood = c("weighted"),
  control = list(...),
  contrasts = NULL,
  ...
)

Arguments

formula

An object of class "formula" which describes the model to be fitted.

data

A data frame containing the variables in the model.

subset

An optional vector specifying a subset of observations to be used.

tau

The quantile to be estimated,

n.plt

The pilot subsample size (the first-step subsample size). These samples will be used to estimate the pilot estimator as well as to estimate the optimal subsampling probability.

n.ssp

The expectation optimal subsample size (the second-step subsample size). For sampling.method = 'withReplacement', n.ssp is exactly the subsample size. For sampling.method = 'poisson', n.ssp is the expectation of subsample size.

B

The number of subsamples for the iterative sampling algorithm. Each subsample contains n.ssp observations. This allows us to estimate the covariance matrix.

boot

If TRUE then perform iterative sampling algorithm and estimate the covariance matrix. If FALSE then only one subsample is returned which contains n.ssp observations.

criterion

The criterion of optimal subsampling probabilities. Choices include optL(default) and uniform.

sampling.method

The sampling method for drawing the optimal subsample. Choices include withReplacement and poisson(default).

likelihood

The type of the maximum likelihood function used to calculate the optimal subsampling estimator. Currently it uses weighted.

control

A list of parameters for controlling the sampling process. Default is list(alpha=0, b=2).

contrasts

An optional list. It specifies how categorical variables are represented in the design matrix. For example, contrasts = list(v1 = 'contr.treatment', v2 = 'contr.sum').

...

A list of parameters which will be passed to quantreg::rq().

Value

ssp.quantreg returns an object of class "ssp.quantreg" containing the following components (some are optional):

model.call

model call

beta.plt

pilot estimator

coefficients

optimal subsample estimator.

cov

covariance matrix of coefficients

index.plt

index of pilot subsample in the full sample

index.ssp

index of optimal subsample in the full sample

N

number of observations in the full sample

subsample.size.expect

expected subsample size

terms

model terms

Details

Additional details... briefly introduce the idea.

References

Wang, H., & Ma, Y. (2021). Optimal subsampling for quantile regression in big data. Biometrika, 108(1), 99-112. doi:10.1093/biomet/asaa043

Examples

#quantile regression
set.seed(1)
N <- 1e4
B <- 5
tau <- 0.75
beta.true <- rep(1, 7)
d <- length(beta.true) - 1
corr  <- 0.5
sigmax  <- matrix(0, d, d)
for (i in 1:d) for (j in 1:d) sigmax[i, j] <- corr^(abs(i-j))
X <- MASS::mvrnorm(N, rep(0, d), sigmax)
err <- rnorm(N, 0, 1) - qnorm(tau)
Y <- beta.true[1] + X %*% beta.true[-1] + 
err * rowMeans(abs(X))
data <- as.data.frame(cbind(Y, X))
colnames(data) <- c("Y", paste("V", 1:ncol(X), sep=""))
formula <- Y ~ .
n.plt <- 200
n.ssp <- 100
optL.results <- ssp.quantreg(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B = B,boot = TRUE,criterion = 'optL',
sampling.method = 'withReplacement',likelihood = 'weighted')
summary(optL.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.quantreg(formula = formula, data = data, tau = tau, n.plt = n.plt, 
#>     n.ssp = n.ssp, B = B, boot = TRUE, criterion = "optL", sampling.method = "withReplacement", 
#>     likelihood = "weighted")
#> 
#> Subsample Size:
#> [1] 500
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept   0.9554     0.0318 30.0568  <0.0001
#> V1          0.9926     0.0567 17.5146  <0.0001
#> V2          0.9761     0.0641 15.2257  <0.0001
#> V3          0.9800     0.0494 19.8297  <0.0001
#> V4          1.0749     0.0361 29.8075  <0.0001
#> V5          0.9890     0.0167 59.3399  <0.0001
#> V6          0.9877     0.0715 13.8179  <0.0001
uni.results <- ssp.quantreg(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B = B,boot = TRUE,criterion = 'uniform',
sampling.method = 'withReplacement', likelihood = 'weighted')
summary(uni.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.quantreg(formula = formula, data = data, tau = tau, n.plt = n.plt, 
#>     n.ssp = n.ssp, B = B, boot = TRUE, criterion = "uniform", 
#>     sampling.method = "withReplacement", likelihood = "weighted")
#> 
#> Subsample Size:
#> [1] 500
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error z value Pr(>|z|)
#> Intercept   1.0404     0.0157 66.4117  <0.0001
#> V1          1.0048     0.0806 12.4606  <0.0001
#> V2          1.0664     0.0687 15.5237  <0.0001
#> V3          0.9842     0.0715 13.7698  <0.0001
#> V4          0.9684     0.0638 15.1784  <0.0001
#> V5          1.0856     0.0378 28.7252  <0.0001
#> V6          0.9097     0.0455 19.9756  <0.0001