Optimal Subsampling for Logistic Regression Model with Rare Events Data

This function fits generalized linear models using ....

ssp.relogit(
  formula,
  data,
  subset = NULL,
  n.plt,
  n.ssp,
  criterion = "optL",
  likelihood = "logOddsCorrection",
  control = list(...),
  contrasts = NULL,
  ...
)

Arguments

formula

An object of class "formula" which describes the model to be fitted.

data

A data frame containing the variables in the model.

subset

An optional vector specifying a subset of observations to be used.

n.plt

The pilot subsample size (the first-step subsample size). These samples will be used to estimate the pilot estimator as well as to estimate the optimal subsampling probability.

n.ssp

The expected optimal subsample size (the second-step subsample size) drawn from those samples with Y=0. All rare events (Y=1) are included in the optimal subsample automatically.

criterion

The criterion of optimal subsampling probabilities. Choices include optA, optL(default), LCC and uniform.

likelihood

The type of the maximum likelihood function used to calculate the optimal subsampling estimator. Choices include weighted and logOddsCorrection(default).

control

A list of parameters for controlling the sampling process. Default is list(alpha=0, b=2).

contrasts

An optional list. It specifies how categorical variables are represented in the design matrix. For example, contrasts = list(v1 = 'contr.treatment', v2 = 'contr.sum').

...

A list of parameters which will be passed to svyglm().

Value

ssp.glm returns an object of class "ssp.glm" containing the following components (some are optional):

model.call

model call

beta.plt

pilot estimator

beta.ssp

optimal subsample estimator.

coefficients

weighted combination of beta.plt and beta.ssp.

cov.ssp

covariance matrix of beta.ssp

cov

covariance matrix of beta.cmb

index.plt

index of pilot subsample in the full sample

index.ssp

index of optimal subsample in the full sample

N

number of observations in the full sample

subsample.size.expect

expected subsample size

terms

model terms

Details

Rare event stands for the number of Y=1 is rare compare to the number of Y=0 in the full sample. When criterion = uniform, it draws (n.plt+n.ssp) subsmples from the full sample with equal sampling probability. When criterion = optA, optL or LCC, observations with Y=1 are preserved and it draw n.ssp subsmples from observations with Y=0.

References

Wang, H., Zhang, A., & Wang, C. (2021). Nonuniform negative sampling and log odds correction with rare events data. Advances in Neural Information Processing Systems, 34, 19847-19859. https://proceedings.neurips.cc/paper_files/paper/2021/hash/a51c896c9cb81ecb5a199d51ac9fc3c5-Abstract.html

Examples

set.seed(1)
N <- 2 * 1e4
beta0 <- c(-5, -rep(0.7, 6))
d <- length(beta0) - 1
X <- matrix(0, N, d)
corr <- 0.5
sigmax <- corr ^ abs(outer(1:d, 1:d, "-"))
sigmax <- sigmax / 4
X <- MASS::mvrnorm(n = N, mu = rep(0, d), Sigma = sigmax)
Y <- rbinom(N, 1, 1 - 1 / (1 + exp(beta0[1] + X %*% beta0[-1])))
print(paste('N: ', N))
#> [1] "N:  20000"
print(paste('sum(Y): ', sum(Y)))
#> [1] "sum(Y):  277"
n.plt <- 200
n.ssp <- 1000
data <- as.data.frame(cbind(Y, X))
colnames(data) <- c("Y", paste("V", 1:ncol(X), sep=""))
formula <- Y ~ .
subsampling.results <- ssp.relogit(formula = formula,
                                     data = data,
                                     n.plt = n.plt,
                                     n.ssp = n.ssp,
                                     criterion = 'optA',
                                     likelihood = 'logOddsCorrection')
summary(subsampling.results)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.relogit(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     criterion = "optA", likelihood = "logOddsCorrection")
#> 
#> Subsample Size:
#>                                 
#> 1       Total Sample Size  20000
#> 2 Expected Subsample Size   1277
#> 3   Actual Subsample Size   1302
#> 4   Unique Subsample Size   1302
#> 5  Expected Subample Rate 6.385%
#> 6    Actual Subample Rate  6.51%
#> 7    Unique Subample Rate  6.51%
#> 
#> Coefficients:
#> 
#>           Estimate Std. Error  z value Pr(>|z|)
#> Intercept  -4.9647     0.0980 -50.6818  <0.0001
#> V1         -0.7250     0.1513  -4.7933  <0.0001
#> V2         -0.9808     0.1633  -6.0076  <0.0001
#> V3         -0.3470     0.1587  -2.1863   0.0288
#> V4         -0.5822     0.1638  -3.5544   0.0004
#> V5         -0.8724     0.1625  -5.3679  <0.0001
#> V6         -0.5414     0.1416  -3.8235   0.0001