Optimal Subsampling Method for Softmax(multinomial logistic) Regression Model

This function fits softmax regression model using ....

ssp.softmax(
  formula,
  data,
  subset,
  n.plt,
  n.ssp,
  criterion = "MSPE",
  sampling.method = "poisson",
  likelihood = "MSCLE",
  constraint = "summation",
  control = list(...),
  contrasts = NULL,
  ...
)

Arguments

formula

An object of class "formula" which describes the model to be fitted.

data

A data frame containing the variables in the model.

subset

An optional vector specifying a subset of observations to be used.

n.plt

The pilot subsample size (the first-step subsample size). These samples will be used to estimate the pilot estimator as well as to estimate the optimal subsampling probability.

n.ssp

The expectation optimal subsample size (the second-step subsample size). For sampling.method = 'withReplacement', n.ssp is exactly the subsample size. For sampling.method = 'poisson', n.ssp is the expectation of subsample size.

criterion

The criterion of optimal subsampling probabilities. Choices include optA, optL, MSPE(default), LUC and uniform.

sampling.method

The sampling method for drawing the optimal subsample. Choices include withReplacement and poisson(default).

likelihood

The type of the maximum likelihood function used to calculate the optimal subsampling estimator. Choices include weighted and MSCLE(default).

constraint

The constraint for identifiability of softmax model. Choices include baseline and summation(default).

control

A list of parameters for controlling the sampling process. Default is list(alpha=0, b=2).

contrasts

An optional list. It specifies how categorical variables are represented in the design matrix. For example, contrasts = list(v1 = 'contr.treatment', v2 = 'contr.sum').

...

A list of parameters which will be passed to nnet::multinom().

Value

ssp.softmax returns an object of class "ssp.softmax" containing the following components (some are optional):

model.call

model call

beta.plt

pilot estimator

beta.ssp

optimal subsample estimator.

coefficients

weighted combination of beta.plt and beta.ssp.

cov.ssp

covariance matrix of beta.ssp

cov

covariance matrix of beta.cmb

index.plt

index of pilot subsample in the full sample

index.ssp

index of optimal subsample in the full sample

N

number of observations in the full sample

subsample.size.expect

expected subsample size

terms

model terms

Details

details TBD

criterion = MSPE stands for optimal subsampling probabilities by minimizing the Mean Squared Prediction Error which is immune to the choice of model constraint.

In control, alpha is the mixture proportions of optimal subsampling probability and uniform sampling probability. b is the parameter controls the upper threshold for optimal subsampling probability.

References

Yao, Y., & Wang, H. (2019). Optimal subsampling for softmax regression. Statistical Papers, 60, 585-599. https://link.springer.com/article/10.1007/s00362-018-01068-6

Yao, Y., Zou, J., & Wang, H. (2023). Optimal poisson subsampling for softmax regression. Journal of Systems Science and Complexity, 36(4), 1609-1625. https://link.springer.com/article/10.1007/s11424-023-1179-z

Yao, Y., Zou, J., & Wang, H. (2023). Model constraints independent optimal subsampling probabilities for softmax regression. Journal of Statistical Planning and Inference, 225, 188-201. doi:10.1016/j.jspi.2022.12.004

Examples

# softmax regression
d <- 3 # dim of covariates
K <- 2 # K + 1 classes
G <- rbind(rep(-1/(K+1), K), diag(K) - 1/(K+1)) %x% diag(d)
N <- 1e4
beta.true.baseline <- cbind(rep(0, d), matrix(-1.5, d, K))
beta.true.summation <- cbind(rep(1, d), 0.5 * matrix(-1, d, K))
set.seed(1)
mu <- rep(0, d)
sigma <- matrix(0.5, nrow = d, ncol = d)
diag(sigma) <- rep(1, d)
X <- MASS::mvrnorm(N, mu, sigma)
prob <- exp( X %*% beta.true.summation)
prob <- prob / rowSums(prob)
Y <- apply(prob, 1, function(row) sample(0:K, size = 1, prob = row))
n.plt <- 500
n.ssp <- 1000
data <- as.data.frame(cbind(Y, X))
colnames(data) <- c("Y", paste("V", 1:ncol(X), sep=""))
head(data)
#>   Y         V1          V2         V3
#> 1 2 -0.3756189 -0.17727086 -0.9816025
#> 2 2  0.2912939  0.60753208 -0.4489936
#> 3 2 -1.0530547  0.02079337 -1.0146024
#> 4 0  0.1854791  2.45385260  1.2682922
#> 5 0  0.8687332  0.21941612 -0.2810234
#> 6 1 -0.8336174 -0.32556141 -0.8505501
formula <- Y ~ . -1
WithRep.MSPE <- ssp.softmax(formula = formula,
 data = data, 
 n.plt = n.plt,
 n.ssp = n.ssp,
 criterion = 'MSPE', 
 sampling.method = 'withReplacement',
 likelihood = 'weighted',
 constraint = 'baseline')
#> # weights:  12 (6 variable)
#> initial  value 549.306144 
#> iter  10 value 367.125951
#> final  value 365.836516 
#> converged
#> # weights:  12 (6 variable)
#> initial  value 11677047.319684 
#> iter  10 value 7149711.444798
#> iter  20 value 6901692.808182
#> final  value 6901684.550943 
#> converged
summary(WithRep.MSPE)
#> Model Summary
#> 
#> 
#> Call:
#> 
#> ssp.softmax(formula = formula, data = data, n.plt = n.plt, n.ssp = n.ssp, 
#>     criterion = "MSPE", sampling.method = "withReplacement", 
#>     likelihood = "weighted", constraint = "baseline")
#> 
#> Subsample Size:
#>                                
#> 1       Total Sample Size 10000
#> 2 Expected Subsample Size  1000
#> 3   Actual Subsample Size  1000
#> 4   Unique Subsample Size   927
#> 5  Expected Subample Rate   10%
#> 6    Actual Subample Rate   10%
#> 7    Unique Subample Rate 9.27%
#> 
#> Coefficients:
#> 
#>         [,1]      [,2]
#> V1 -1.455068 -1.406786
#> V2 -1.647116 -1.492484
#> V3 -1.248385 -1.435806
#> 
#> Std. Errors:
#> 
#>          [,1]       [,2]
#> V1 0.08803001 0.08917739
#> V2 0.08516142 0.08456003
#> V3 0.08726432 0.08841460